requirements.txt
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
openai==1.3.7
python-dotenv==1.0.0
streamlit==1.28.1
pandas==2.1.3
beautifulsoup4==4.12.2
requests==2.31.0
urllib3==2.1.0
python-dateutil==2.8.2
protobuf==4.25.8
grpcio-status==1.48.2
httpx==0.24.1 